Renamed module from aci to container_instance
